More informative per-view error logging
Now, it's easier to get the index of the pose that's responsible for
bad reprojection errors, as indexing eliminates the need for manual
counting.

A future improvement could be re-designing the naming scheme of the
capture directories to be index-based instead of time-based, so
manual counting of the directories is unnecessary.