Optimize capture speed of Photoneo MotionCam
We can safely decrease the wait time to 0 and take captures
continuously if we add a lag at the beginning of the captures. Still,
a limitation exists: the image acquisition runs independently of the
project and capture module. This independence is not a problem if the
FPS is sufficiently high, but becomes a problem once the FPS is low
(e.g., MotionCam's Scanner Mode's FPS of 3). Theoretically, this
low FPS can cause major issues in synchronization, but in practice,
the captures always look synchronized.

The next step is to make procamcalib capture more robust while
keeping the same speed.