Sync projection and image acquisition
Now, the projection and image acquisitions are totally synced by
means of system clocks. These clocks make sure that the image
is accepted as part of an image block only if its acquisition occurs
after the projection.

With this fix, the inelegant 500ms initial lag becomes unnecessary.
The speed is also reasonably good although not exactly as fast as
without the synchronization.